preparing environment for a more systemize approach: created work and version control flow protocols, deleted all experiments from deploy_3_6, redesigned "general considerations", updated project status
