+++ general considerations/system configuration/version control/version control protocol
@@ -1,0 +1,8 @@
+fiecare branch se creeaza si se mergeuieste in develop-seriously
+in master se face merge doar cand poate fi facut deploy ceea ce se va si face dupa fiecare merge
+in fiecare commit se explica exact ce s-a facut
+patternul pt brancheul din develop-seriously e urmatorul: urmatorul_deploy/ce_se_face
+dupa deploy se adauga fisierele( cu un scurt rezumat si concluzii)
+si modele ce trebuiesc salvate in history si se commit din master cu urmatorul pattern nume_deploy_incorporated_in_history
+modificarile ce tin de sistem se fac pe un branch cu nume sugestiv din develop-seriously care se propaga pe toate branchurile pe care se lucreaza
+schimbarile pe fisierul de project status se commit impreuna cu alt task, nu au nevoie de  branch separat
