--- models/modelsMountainCar/history/deploy_3_6/online/details
@@ -1,16 +1,0 @@
- model.compile(loss=tf.keras.losses.Huber(), optimizer=Adam(lr=self.learning_rate)) un dezastru -> nu le pot deschide
-
-stack_depth = 4
-
-in general rezultate foarte proste->o convers foarte greu si nici o solutie nu e validata la test :(
-
-num_pic_from buffer = 32
-minimum_samples before started the training: 150000
-max size buffer to pick up from samples for training :200000
-
-hidden layers in network:
-
-2 -> Conv2D
-1-> Dense
-
-Q values la antrenare nu sunt setate pe 0 si se face doar updateul pe  best action
