--- models/modelsMountainCar/history/deploy_6_6/details
@@ -1,1 +1,0 @@
-increase the samples on which the training is done from 32 -> 50
